third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/services-found-with-uuid.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(function() {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      gattServer.getPrimaryServices(heart_rate.alias),
      gattServer.getPrimaryServices(heart_rate.name),
      gattServer.getPrimaryServices(heart_rate.uuid)]))
    .then(services_arrays => {
      services_arrays.forEach(services => {
        assert_equals(services.length, 2);
        assert_equals(services[0].uuid,
                      BluetoothUUID.getService('heart_rate'));
        assert_equals(services[1].uuid,
                      BluetoothUUID.getService('heart_rate'));
        assert_true(services[0].isPrimary);
        assert_true(services[1].isPrimary);
      });
    });
}, 'Request for services. Should return right services');
</script>
